requirements.txt
langchain
tiktoken
pinecone-client
pypdf
openai
langchain
pandas
numpy
python-dotenv
Pypdf2
pypdf
pdf2image
pytesseract
langchain-google-genai
pinecone
together
